Move IsLayoutInUse method to LayoutHelper class
